move UncompressedBufferPageMapping behind impl
