+++ tt_metal/impl/buffers/buffer_page_mapping.cpp
@@ -2,6 +2,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
+#include "impl/buffers/buffer_page_mapping.hpp"
 #include <tt-metalium/buffer_page_mapping.hpp>
 
 namespace tt::tt_metal {
